requirements.txt
requests
py-cord
pyyaml
ics
arrow